pyproject.toml
[tool.poetry]
name = "ai-call-agent"
version = "0.1.0"
description = "OpenAI Realtime API + Twilio Voice Integration"
authors = ["Rehan Khan"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohappyeyeballs = "2.4.4"
aiohttp = "3.11.10"
aiohttp-retry = "2.8.3"
aiosignal = "1.3.1"
annotated-types = "0.7.0"
anyio = "4.7.0"
attrs = "24.2.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
click = "8.1.7"
distro = "1.9.0"
fastapi = "0.115.6"
frozenlist = "1.5.0"
h11 = "0.14.0"
httpcore = "1.0.7"
httpx = "0.28.1"
idna = "3.10"
jiter = "0.8.0"
multidict = "6.1.0"
observable = "0.3.2"
openai = "1.57.1"
propcache = "0.2.1"
pydantic = "2.10.3"
pydantic_core = "2.27.1"
PyJWT = "2.10.1"
python-dotenv = "1.0.1"
requests = "2.32.3"
ringcentral = "0.9.1"
sniffio = "1.3.1"
starlette = "0.41.3"
tqdm = "4.67.1"
twilio = "9.3.8"
typing_extensions = "4.12.2"
urllib3 = "2.2.3"
uvicorn = "0.32.1"
websockets = "12.0"
yarl = "1.18.3"
structlog = "25.4.0"
google-api-python-client = "2.129.0"
google-auth = "2.29.0"
google-auth-httplib2 = "0.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
